去掉显示用户信息页面
